# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/users/update-user": {
      "put": {
        "operationId": "UsersController_updateUser",
        "parameters": [
          {
            "name": "auth_token",
            "in": "header",
            "description": "Token xác thực của người dùng",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Thong tin cap nhat",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/users/get-user-info": {
      "post": {
        "operationId": "UsersController_getUserInfo",
        "parameters": [
          {
            "name": "auth_token",
            "in": "header",
            "description": "Token xác thực của người dùng",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/users/remove-user/{id}": {
      "delete": {
        "operationId": "UsersController_deleteUser",
        "parameters": [
          {
            "name": "auth_token",
            "in": "header",
            "description": "Token xác thực của người dùng",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "required": false,
            "in": "path",
            "description": "remove user id by ADMIN",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/users/get-user-list/{page}/{size}/{key}": {
      "get": {
        "operationId": "UsersController_getListUserPagination",
        "parameters": [
          {
            "name": "auth_token",
            "in": "header",
            "description": "Token xác thực của người dùng",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "path",
            "description": "Search",
            "schema": {}
          },
          {
            "name": "size",
            "required": false,
            "in": "path",
            "description": "item per page",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "path",
            "description": "page number",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Thong tin dang nhap",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/sign-up": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Thong tin dang ky",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/product-upc": {
      "get": {
        "operationId": "ProductUpcController_getAllProductUpcs",
        "parameters": [
          {
            "name": "auth_token",
            "in": "header",
            "description": "Token xác thực của người dùng",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Product-upc"
        ]
      }
    },
    "/product-upc/create": {
      "post": {
        "operationId": "ProductUpcController_createProduct",
        "parameters": [
          {
            "name": "auth_token",
            "in": "header",
            "description": "Token xác thực của người dùng",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Dữ liệu tạo sản phẩm",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Tên sản phẩm"
                  },
                  "description": {
                    "type": "string",
                    "description": "Mô tả sản phẩm"
                  },
                  "upc": {
                    "type": "string",
                    "description": "Mã UPC"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Ảnh sản phẩm"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Product-upc"
        ]
      }
    }
  },
  "info": {
    "title": "The point saver API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "user_name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zip": {
            "type": "number"
          },
          "country": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "user_name",
          "phone",
          "password",
          "address",
          "city",
          "state",
          "zip",
          "country"
        ]
      },
      "loginDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "email"
          },
          "password": {
            "type": "string",
            "description": "password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "user_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zip": {
            "type": "number"
          },
          "country": {
            "type": "string"
          }
        },
        "required": [
          "user_name",
          "email",
          "phone",
          "password",
          "address",
          "city",
          "state",
          "zip",
          "country"
        ]
      }
    }
  }
}